+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## V1.1.0
+* abstract class [BaseEntity](/src/main/java/it/sirziphonprojects/hibernateentitymodel/entity/BaseEntity.java) added parameter id  an implementation of method ```public boolean equals(Map<String, Object> keysMap)```
+* interface [KeyMapper](/src/main/java/it/sirziphonprojects/hibernateentitymodel/entity/KeyMapper.java) added method ```public boolean equals(Map<String, Object> keysMap)```
+
 ## V1.0.1
 * fixed minor issues